requirements.txt
numpy==1.24.3
flask==2.0.1
werkzeug==2.0.1
pytesseract==0.3.8
pdf2image==1.16.0
pillow==8.3.1
requests==2.26.0
redis==3.5.3
rq==1.10.0
python-dotenv==0.19.0
gunicorn==20.1.0
opencv-python-headless==4.5.3.56
python-stdnum